+++ src/Nethermind/Nethermind.JsonRpc/JsonRpcConfig.cs
@@ -68,5 +68,5 @@ public class JsonRpcConfig : IJsonRpcConfig
     public int EstimateErrorMargin { get; set; } = 150;
     public string[] CorsOrigins { get; set; } = ["*"];
     public int WebSocketsProcessingConcurrency { get; set; } = 1;
-    public long IpcProcessingConcurrency { get; set; } = 1;
+    public int IpcProcessingConcurrency { get; set; } = 1;
 };
